Rewrote how attributes from a node and it's groups are looked up and added the ability to concatinate values for these attributes from mulitple sources by adding "+=NEXTRECORD" to the values in the data.
For example, if bootparams.addkcmdline for the node is "test1+=NEXTRECORD" and bootparams.addkcmdline for the group it is in is "test2", the returned values from nodels for bootparams.addkcmdline will be "test1 test2".

Not that this stops after it finds a single occurance of a value for this attriute without the "+=NEXTRECORD".  If all values from the groups are wanted, each should have "+=NEXTRECORD" at the end of the entry.



git-svn-id: http://xcat.svn.sourceforge.net/svnroot/xcat/xcat-core/trunk@6840 35838ccf-6b3d-0410-a007-ed7a3c9b06b9